# docs/swagger.yaml
basePath: /
definitions:
  handlers.DatosProcesadosResponse:
    properties:
      productos:
        items:
          additionalProperties: true
          type: object
        type: array
      sucursales:
        items:
          additionalProperties: true
          type: object
        type: array
      ventas:
        items:
          additionalProperties: true
          type: object
        type: array
    type: object
  handlers.DatosProcesamientoRequest:
    type: object
  handlers.ErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message:
        example: JSON inválido
        type: string
      time:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  handlers.ProcesamientoResponse:
    properties:
      message:
        example: Datos procesados y depurados y reporte generado
        type: string
      status:
        example: Proceso completado exitosamente
        type: string
      time:
        example: "2024-01-15T10:30:00Z"
        type: string
    type: object
  handlers.ReporteResponse:
    properties:
      Datos:
        additionalProperties: true
        type: object
      Estado:
        example: completado
        type: string
      FechaFin:
        example: "2024-01-15T23:59:59Z"
        type: string
      FechaInicio:
        example: "2024-01-01T00:00:00Z"
        type: string
      Filtros:
        example:
        - '["fecha_inicio"'
        - '"fecha_fin"]'
        items:
          type: string
        type: array
      GeneradoEn:
        example: "2024-01-15T12:00:00Z"
        type: string
      ID:
        example: 1
        type: integer
      Resumen:
        example: Reporte de ventas del período
        type: string
      Tipo:
        example: ventas
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API RESTful con arquitectura dirigida por eventos para procesamiento,
    depuración y consulta de datos de productos, ventas y sucursales
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Sistema de Procesamiento de Datos API
  version: "1.0"
paths:
  /api/datos-procesados:
    get:
      description: Obtiene los datos procesados y depurados almacenados en memoria
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.DatosProcesadosResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Consultar datos procesados
      tags:
      - procesamiento
  /api/procesar:
    post:
      consumes:
      - application/json
      description: Recibe datos crudos, los procesa y depura, y dispara eventos de
        notificación
      parameters:
      - description: Datos crudos a procesar
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.DatosProcesamientoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ProcesamientoResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Procesar datos crudos
      tags:
      - procesamiento
  /api/reporte:
    get:
      description: Obtiene el último reporte generado usando el patrón Builder
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ReporteResponse'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Obtener último reporte
      tags:
      - procesamiento
schemes:
- http
swagger: "2.0"